requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.5
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.1
drf-nested-routers==0.94.2
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
